requirements.txt
asgiref==3.8.1
automium==0.2.6
automium_web==0.1.1
axios==0.4.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
legacy==0.1.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.6
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.3
rich==13.9.4
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
